requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flask-marshmallow==0.14.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
mysql-connector-python==8.0.27
protobuf==3.19.1
PyJWT==2.3.0
six==1.16.0
Werkzeug==2.0.2